+++ packages/common/src/logger.ts
@@ -301,4 +301,3 @@ export class NativeLogger implements Logger {
 		}
 	}
 }
-