+++ packages/provider/src/api/ja4Middleware.ts
@@ -16,7 +16,7 @@ import { createHash } from "node:crypto";
 import type { IncomingHttpHeaders } from "node:http";
 import { Readable } from "node:stream";
 import { handleErrors } from "@prosopo/api-express-router";
-import { type Logger, getLogger } from "@prosopo/common";
+import { type Logger, getLoggerDefault } from "@prosopo/common";
 import type { ProviderEnvironment } from "@prosopo/types-env";
 import type { NextFunction, Request, Response } from "express";
 import { readTlsClientHello } from "read-tls-client-hello";
@@ -24,7 +24,7 @@ import { readTlsClientHello } from "read-tls-client-hello";
 export const DEFAULT_JA4 = "ja4";
 
 export const getJA4 = async (headers: IncomingHttpHeaders, logger?: Logger) => {
-	logger = logger || getLogger('info', import.meta.url);
+	logger = logger || getLoggerDefault();
 
 	// Default JA4+ fingerprint for development
 	if (process.env.NODE_ENV === "development") {
@@ -43,23 +43,18 @@ export const getJA4 = async (headers: IncomingHttpHeaders, logger?: Logger) => {
 		const clientHelloBuffer = Buffer.from(xTlsClientHello, "base64");
 
 		// Debug: Check first few bytes
-		logger.debug(() => ({
-			data: { data: clientHelloBuffer.subarray(0, 5).toString("hex") },
-			msg: "ClientHello First Bytes"
-		}));
+		logger.debug(
+			"ClientHello First Bytes:",
+			clientHelloBuffer.subarray(0, 5).toString("hex"),
+		);
 
 		// Check first byte after the initial 5
 		if (clientHelloBuffer[5] !== 0x01) {
-			logger.warn(() => ({
-				msg: "Invalid ClientHello message: First byte is not 0x01"
-			}));
+			logger.debug("Invalid ClientHello message: First byte is not 0x01");
 			return { ja4PlusFingerprint: DEFAULT_JA4 };
 		}
 
-		logger.debug(() => ({
-			data: { xTlsVersion },
-			msg: "Headers TLS Version"
-		}));
+		logger.debug("Headers TLS Version:", xTlsVersion);
 
 		//`tls1.3` becomes `13` etc.
 		const tlsVersion = xTlsVersion.replace(/(tls)|\./g, "");
@@ -132,10 +127,8 @@ export const getJA4 = async (headers: IncomingHttpHeaders, logger?: Logger) => {
 		const ja4PlusFingerprint = `${transport}${tlsVersion}${sniIndicator}${cipherCount}${extensionCount}${alpnLabel}_${cipherHash}_${extensionHash}`;
 
 		return { ja4PlusFingerprint };
-	} catch (err) {
-		logger.error(() => {
-			return { err, msg: "Error generating JA4+ fingerprint" };
-		});
+	} catch (e) {
+		logger.error("Error generating JA4+ fingerprint:", e);
 		return { ja4PlusFingerprint: DEFAULT_JA4 };
 	}
 };
@@ -143,6 +136,7 @@ export const getJA4 = async (headers: IncomingHttpHeaders, logger?: Logger) => {
 export const ja4Middleware = (env: ProviderEnvironment) => {
 	return async (req: Request, res: Response, next: NextFunction) => {
 		try {
+			req.logger.debug({ url: req.url });
 			const ja4 = await getJA4(req.headers, req.logger);
 
 			req.ja4 = ja4.ja4PlusFingerprint || "";