+++ packages/provider/src/tasks/powCaptcha/powTasks.ts
@@ -11,14 +11,11 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import type { KeyringPair } from "@polkadot/keyring/types";
+
 import { stringToHex, u8aToHex } from "@polkadot/util";
-import {
-	ProsopoApiError,
-	ProsopoEnvError,
-	getLogger,
-} from "@prosopo/common";
+import { ProsopoApiError, ProsopoEnvError } from "@prosopo/common";
 import type { Logger } from "@prosopo/common";
+import type { KeyringPair } from "@prosopo/types";
 import {
 	ApiParams,
 	type CaptchaResult,
@@ -27,7 +24,6 @@ import {
 	POW_SEPARATOR,
 	type PoWCaptcha,
 	type PoWChallengeId,
-	type ProsopoConfigOutput,
 	type RequestHeaders,
 } from "@prosopo/types";
 import type { IProviderDatabase } from "@prosopo/types-database";
@@ -123,7 +119,7 @@ export class PowCaptchaManager extends CaptchaManager {
 			await this.db.getPowCaptchaRecordByChallenge(challenge);
 
 		if (!challengeRecord) {
-			this.logger.debug(() => ({ msg: "No record of this challenge" }));
+			this.logger.debug("No record of this challenge");
 			// no record of this challenge
 			return false;
 		}
@@ -181,24 +177,20 @@ export class PowCaptchaManager extends CaptchaManager {
 			await this.db.getPowCaptchaRecordByChallenge(challenge);
 
 		if (!challengeRecord) {
-			this.logger.debug(() => ({ data: { challenge }, msg: "No record of this challenge" }));
+			this.logger.debug(`No record of this challenge: ${challenge}`);
 			return { verified: false };
 		}
 
 		if (ip) {
 			const ipV4Address = getIPAddress(ip);
 			if (!ipV4Address) {
-				this.logger.debug(() => ({ data: { ip }, msg: "Invalid IP address" }));
+				this.logger.debug(`Invalid IP address: ${ip}`);
 				return { verified: false };
 			}
 			if (challengeRecord.ipAddress !== ipV4Address.bigInt()) {
-				this.logger.debug(() => ({
-					data: {
-						expected: getIPAddressFromBigInt(challengeRecord.ipAddress).address,
-						actual: ip
-					},
-					msg: "IP address mismatch"
-				}));
+				this.logger.debug(
+					`IP address mismatch: ${getIPAddressFromBigInt(challengeRecord.ipAddress).address} !== ${ip}`,
+				);
 				return { verified: false };
 			}
 		}
@@ -239,12 +231,10 @@ export class PowCaptchaManager extends CaptchaManager {
 			);
 			if (tokenRecord) {
 				score = computeFrictionlessScore(tokenRecord?.scoreComponents);
-				this.logger.info(() => ({
-					data: {
-						tscoreComponents: tokenRecord?.scoreComponents,
-						score: score,
-					}
-				}));
+				this.logger.info({
+					tscoreComponents: tokenRecord?.scoreComponents,
+					score: score,
+				});
 			}
 		}