+++ packages/common/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @prosopo/common
 
+## 2.7.0
+### Minor Changes
+
+- 8f0644a: Taking required functions from polkadot/keyring and polkadot/util-crypto in-house and removing WASM dependencies. Adding @scure JS-based sr25519 function instead.
+
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/util-crypto@13.4.0
+
 ## 2.6.1
 
 ### Patch Changes