+++ packages/load-balancer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prosopo/load-balancer
 
+## 2.6.5
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/common@2.7.0
+  - @prosopo/types@2.8.0
+
 ## 2.6.4
 
 ### Patch Changes