+++ packages/procaptcha-bundle/CHANGELOG.md
@@ -1,5 +1,69 @@
 # @prosopo/procaptcha-bundle
 
+## 2.7.0
+### Minor Changes
+
+- 8f0644a: Taking required functions from polkadot/keyring and polkadot/util-crypto in-house and removing WASM dependencies. Adding @scure JS-based sr25519 function instead.
+
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/procaptcha-pow@2.7.0
+  - @prosopo/util-crypto@13.4.0
+  - @prosopo/common@2.7.0
+  - @prosopo/types@2.8.0
+  - @prosopo/procaptcha-react@2.6.14
+  - @prosopo/procaptcha-frictionless@2.6.14
+  - @prosopo/dotenv@2.6.2
+
+## 2.6.13
+### Patch Changes
+
+- Updated dependencies [cf26d7e]
+  - @prosopo/procaptcha-frictionless@2.6.13
+  - @prosopo/procaptcha-pow@2.6.13
+  - @prosopo/procaptcha-react@2.6.13
+
+## 2.6.12
+
+### Patch Changes
+
+- ea38a1c: lint
+- Updated dependencies [ea38a1c]
+  - @prosopo/procaptcha-frictionless@2.6.12
+  - @prosopo/procaptcha-react@2.6.12
+  - @prosopo/procaptcha-pow@2.6.12
+
+## 2.6.11
+
+### Patch Changes
+
+- b2ae723: lint
+- Updated dependencies [b2ae723]
+  - @prosopo/procaptcha-frictionless@2.6.11
+  - @prosopo/procaptcha-react@2.6.11
+  - @prosopo/procaptcha-pow@2.6.11
+
+## 2.6.10
+
+### Patch Changes
+
+- d17c67f: lint
+- Updated dependencies [d17c67f]
+  - @prosopo/procaptcha-frictionless@2.6.10
+  - @prosopo/procaptcha-react@2.6.10
+  - @prosopo/procaptcha-pow@2.6.10
+
+## 2.6.9
+
+### Patch Changes
+
+- 0d194f2: lint
+- Updated dependencies [0d194f2]
+  - @prosopo/procaptcha-frictionless@2.6.9
+  - @prosopo/procaptcha-react@2.6.9
+  - @prosopo/procaptcha-pow@2.6.9
+
 ## 2.6.8
 
 ### Patch Changes