+++ packages/procaptcha-common/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @prosopo/procaptcha-common
 
+## 2.7.0
+### Minor Changes
+
+- 8f0644a: Taking required functions from polkadot/keyring and polkadot/util-crypto in-house and removing WASM dependencies. Adding @scure JS-based sr25519 function instead.
+
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/account@2.7.0
+  - @prosopo/common@2.7.0
+  - @prosopo/types@2.8.0
+  - @prosopo/load-balancer@2.6.5
+
+## 2.6.11
+
+### Patch Changes
+
+- ea38a1c: lint
+
+## 2.6.10
+
+### Patch Changes
+
+- b2ae723: lint
+
+## 2.6.9
+
+### Patch Changes
+
+- d17c67f: lint
+
+## 2.6.8
+
+### Patch Changes
+
+- 0d194f2: lint
+
 ## 2.6.7
 
 ### Patch Changes