+++ packages/procaptcha/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @prosopo/procaptcha
 
+## 2.7.0
+### Minor Changes
+
+- 8f0644a: Taking required functions from polkadot/keyring and polkadot/util-crypto in-house and removing WASM dependencies. Adding @scure JS-based sr25519 function instead.
+
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/procaptcha-common@2.7.0
+  - @prosopo/util-crypto@13.4.0
+  - @prosopo/datasets@2.7.0
+  - @prosopo/account@2.7.0
+  - @prosopo/common@2.7.0
+  - @prosopo/types@2.8.0
+  - @prosopo/load-balancer@2.6.5
+  - @prosopo/api@2.6.5
+
+## 2.6.12
+
+### Patch Changes
+
+- Updated dependencies [ea38a1c]
+  - @prosopo/procaptcha-common@2.6.11
+  - @prosopo/datasets@2.6.12
+
+## 2.6.11
+
+### Patch Changes
+
+- Updated dependencies [b2ae723]
+  - @prosopo/procaptcha-common@2.6.10
+  - @prosopo/datasets@2.6.11
+
+## 2.6.10
+
+### Patch Changes
+
+- Updated dependencies [d17c67f]
+  - @prosopo/procaptcha-common@2.6.9
+  - @prosopo/datasets@2.6.10
+
+## 2.6.9
+
+### Patch Changes
+
+- Updated dependencies [0d194f2]
+  - @prosopo/procaptcha-common@2.6.8
+  - @prosopo/datasets@2.6.9
+
 ## 2.6.8
 
 ### Patch Changes