+++ packages/server/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @prosopo/server
 
+## 2.7.0
+### Minor Changes
+
+- 8f0644a: Taking required functions from polkadot/keyring and polkadot/util-crypto in-house and removing WASM dependencies. Adding @scure JS-based sr25519 function instead.
+
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/keyring@2.7.0
+  - @prosopo/common@2.7.0
+  - @prosopo/types@2.8.0
+  - @prosopo/load-balancer@2.6.5
+  - @prosopo/api@2.6.5
+
 ## 2.6.4
 
 ### Patch Changes